Use urdfdom_headers dependency key. (#6)
While the upstream package is called urdfdom-headers, the project-level
name is urdfdm_headers and ament_tools and cmake both expect to find
urdfdom_headers. Short of inserting hacks into those tools or changing
additional sources there isn't an easy or straightforward way to use the
dependency key urdfdom-headers and have it continue to work the way
source builds currently work on ROS 2 CI.

This project was depending on `urdfdom-headers` which I think means that
there was no direct rule guaranteeing that `urdfdom_headers` would be
available except perhaps as a dependency of another dependency, urdfdom.

This changes the key to depend on `urdfdom_headers` directly and notes
when we should be able to revert to using a rosdep key conventional
name.